requirements.txt
Django>=3.2,<4.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
gunicorn
django-cors-headers